Refactor coconut e2e test for reuse
