+++ common/nymcoconut/src/tests/mod.rs
@@ -1,1 +1,3 @@
+#[cfg(test)]
 mod e2e;
+pub mod helpers;
